Implement branch-and-bound for part 2
I'm not doing the greedy method anymore - I have test cases that
will catch this (finally). The method that most people use on
Reddit is to just let the human take as much as possible within 26
turns, then give the remaining valves to the elephant. This is not
in the spirit of the question, and I've since implemented a search
that uses brute force + bnb

- Added an extra turn for the canonical graph building. For some
reason, input5 doesn't work properly and requires an extra turn to
build the canonical graph... adding +1 fixed it

- Remove max_flow as I no longer use it

- Remove some old variables

- Implement ignore_set for part 2

- Add subset utility functions